Change gotonpc to gotoname
